% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{print.pfsqca_params}
\alias{print.pfsqca_params}
\title{Print method for pfsqca_params}
\usage{
\method{print}{pfsqca_params}(x, ...)
}
\arguments{
\item{x}{A pfsqca_params object.}

\item{...}{Additional arguments (ignored).}
}
\value{
Invisibly returns x.
}
\description{
Print method for pfsqca_params
}
