% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/necessity.R
\name{necessity_panel_diagnostics}
\alias{necessity_panel_diagnostics}
\title{Panel Diagnostics for Necessity}
\usage{
necessity_panel_diagnostics(
  data,
  outcome,
  conditions,
  id_var = "MSA",
  time_var = "year"
)
}
\arguments{
\item{data}{A data frame with calibrated fuzzy-set variables.}

\item{outcome}{Character. Name of the outcome variable.}

\item{conditions}{Character vector. Names of condition variables.}

\item{id_var}{Character. Name of the case identifier variable. Default is "MSA".}

\item{time_var}{Character. Name of the time variable. Default is "year".}
}
\value{
A list with two data frames:
\describe{
\item{between}{Necessity metrics by time period (BECONS_N, BECOV_N)}
\item{within}{Necessity metrics by case (WICONS_N, WICOV_N)}
}
}
\description{
Calculates between-period and within-case necessity metrics
to assess stability of necessity relations across time and units.
}
\examples{
\dontrun{
diag <- necessity_panel_diagnostics(data_cal, "outcome", c("income", "patents"))
diag$between  # by year
diag$within   # by case
}

}
