% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robustness.R
\name{necessity_robustness}
\alias{necessity_robustness}
\title{Robustness Test for Necessity (Random Deletion)}
\usage{
necessity_robustness(data, outcome, conditions, params = NULL)
}
\arguments{
\item{data}{A data frame with calibrated fuzzy-set variables.}

\item{outcome}{Character. Name of the outcome variable.}

\item{conditions}{Character vector. Names of condition variables.}

\item{params}{A pfsqca_params object. If NULL, uses default parameters.}
}
\value{
A data frame with:
\describe{
\item{condition}{The condition tested}
\item{robust_share}{Proportion of iterations where condition was necessary}
\item{robust_nec_flag}{Whether robust_share >= params$robust_nec_cut}
}
}
\description{
Tests robustness of necessity findings by randomly deleting
observations and re-testing necessity multiple times.
}
\examples{
\dontrun{
params <- pfsqca_params(robustness_B = 100)
rob_nec <- necessity_robustness(data_cal, "outcome", conditions, params)
rob_nec  # see which conditions are robustly necessary
}

}
