% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sufficiency.R
\name{panel_metrics}
\alias{panel_metrics}
\title{Panel Metrics for Sufficiency}
\usage{
panel_metrics(data, outcome, sol_terms, id_var = "MSA", time_var = "year")
}
\arguments{
\item{data}{A data frame with calibrated fuzzy-set variables.}

\item{outcome}{Character. Name of the outcome variable.}

\item{sol_terms}{Character vector. Solution terms from sufficiency analysis.}

\item{id_var}{Character. Name of the case identifier variable. Default is "MSA".}

\item{time_var}{Character. Name of the time variable. Default is "year".}
}
\value{
A list with three data frames:
\describe{
\item{pooled}{POCONS and POCOV for each term and overall solution}
\item{between}{BECONS and BECOV by time period}
\item{within}{WICONS and WICOV by case}
}
}
\description{
Calculates pooled, between-period, and within-case consistency and coverage
metrics for sufficient configurations.
}
\examples{
\dontrun{
suf <- sufficiency_analysis(data_cal, "outcome", conditions, params)
pm <- panel_metrics(data_cal, "outcome", suf$terms)
pm$pooled   # overall metrics
pm$between  # by year
pm$within   # by case
}

}
